+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### [Added]
 
+* StoreEntityResolver: possibilité de récupérer une liste d'entités ou d’informations sur une entité #85
 * ajout classe Permission : gère les permissions sur les offres #93
 * ajout classe PermissionAction : gère les permissions sur les offres depuis le workflow #94